pyproject.toml
# Agent Forge V2 - Project Configuration

[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-forge-v2"
version = "1.0.0"
description = "8-phase AI agent creation pipeline with 25M parameter models"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Agent Forge Team"}
]
keywords = ["ai", "machine-learning", "pytorch", "transformers", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.0.0",
    "transformers>=4.35.0",
    "peft>=0.7.0",
    "wandb>=0.16.0",
    "pyyaml>=6.0",
    "numpy>=1.24.0",
    "tqdm>=4.65.0",
    "safetensors>=0.4.0",  # ISS-004: Secure checkpoint format (no pickle)
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.3.0",
    "black>=23.12.0",
    "isort>=5.13.0",
    "flake8>=7.0.0",
    "pylint>=3.0.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]
ui = [
    "streamlit>=1.28.0",
    "pandas>=2.0.0",
    "psutil>=5.9.0",
]

[project.urls]
Homepage = "https://github.com/agent-forge/agent-forge-v2"
Documentation = "https://github.com/agent-forge/agent-forge-v2/docs"
Repository = "https://github.com/agent-forge/agent-forge-v2"

# Black configuration
[tool.black]
line-length = 100
target-version = ['py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | storage
  | wandb
)/
'''

# isort configuration
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_gitignore = true

# pytest configuration
[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-report=xml",
    "--cov-fail-under=90",  # Require ≥90% coverage
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "gpu: marks tests requiring GPU",
    "phase1: marks tests for Phase 1",
    "phase2: marks tests for Phase 2",
    "phase3: marks tests for Phase 3",
    "phase4: marks tests for Phase 4",
    "merge_technique: marks tests for Phase 2 merge techniques",
    "fitness: marks tests for Phase 2 fitness evaluation",
    "genetic_ops: marks tests for Phase 2 genetic operations",
    "population: marks tests for Phase 2 population management",
    "handoff: marks tests for phase handoff validation",
]

# Coverage configuration
[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/__init__.py",
    "*/setup.py",
]
branch = true

[tool.coverage.report]
precision = 2
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

# mypy configuration
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = false
disallow_subclassing_any = false
disallow_untyped_calls = false
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
strict_optional = true

[[tool.mypy.overrides]]
module = [
    "torch.*",
    "transformers.*",
    "peft.*",
    "wandb.*",
    "streamlit.*",
]
ignore_missing_imports = true

# Pylint configuration
[tool.pylint.main]
max-line-length = 100
disable = [
    "C0103",  # Invalid name
    "C0114",  # Missing module docstring
    "C0115",  # Missing class docstring
    "C0116",  # Missing function docstring
    "R0903",  # Too few public methods
    "R0913",  # Too many arguments
    "W0212",  # Protected access
]
good-names = ["i", "j", "k", "x", "y", "z", "f", "lr", "df", "db", "id"]

[tool.pylint.design]
max-args = 7
max-locals = 15
max-branches = 12
max-statements = 60  # Aligned with NASA POT10
